src/main/java/skypro/hw28/service/DepartmentService.java
package skypro.hw28.service;

import org.springframework.stereotype.Service;
import skypro.hw28.Employee;
import skypro.hw28.exceptions.EmployeeNotFoundException;

import java.util.Comparator;
import java.util.List;
import java.util.Map;
import java.util.stream.Collectors;

@Service
public class DepartmentService {

    private final EmployeeService employeeService;

    public DepartmentService(EmployeeService employeeService) {
        this.employeeService = employeeService;
    }

    public Employee findEmployeeWithMaxSalaryFromDepartment(int department) {
    return employeeService.getAll().stream()
            .filter(epmloyee -> epmloyee.getDepartment() == department)
            .max(Comparator.comparingDouble(Employee::getSalary))
            .orElseThrow(() -> new EmployeeNotFoundException());
    }

    public Employee findEmployeeWithMinSalaryFromDepartment(int department) {
        return employeeService.getAll().stream()
                .filter(epmloyee -> epmloyee.getDepartment() == department)
                .min(Comparator.comparingDouble(Employee::getSalary))
                .orElseThrow(() -> new EmployeeNotFoundException());
    }

    public List<Employee> findEmployeesFromDepartment(int department) {
        return employeeService.getAll().stream()
                .filter(epmloyee -> epmloyee.getDepartment() == department)
                .collect(Collectors.toList());
    }

    public Map<Integer, List<Employee>> findEmployees() {
        return employeeService.getAll().stream()
                .collect(Collectors.groupingBy(Employee::getDepartment));
    }
}
